requirement.txt
pytorch==1.8.0
transformers==4.2.0.dev0
attrdict==2.0.1
attrs==19.1.0
boto3==1.9.199
botocore==1.12.199
certifi==2019.6.16
chardet==3.0.4
comet-git-pure==0.19.11
comet-ml==2.0.5
configobj==5.0.6
cycler==0.10.0
docutils==0.14
everett==1.0.2
editdistance==0.3.1
idna==2.8
jmespath==0.9.4
joblib==0.13.2
jsonschema==3.0.1
kiwisolver==1.1.0
matplotlib==3.1.1
netifaces==0.10.9
nltk==3.4.5
numpy==1.17.0
nvidia-ml-py3==7.352.0
pandas==0.24.2
Pillow==6.1.0
pyparsing==2.4.2
pyrouge==0.1.3
pyrsistent==0.15.4
python-dateutil==2.8.0
regex==2019.6.8
requests==2.22.0
s3transfer==0.2.1
scipy==1.3.0
six==1.12.0
torch==1.3.0
tqdm==4.32.2
urllib3==1.25.3
websocket-client==0.56.0
wurlitzer==1.0.3
gensim==3.8.0
zss==1.2.0
h5py==2.10.0
subword-nmt
ipdb